+++ app/src/main/java/com/vu/androidbasicapp/ProfileFragment.kt
@@ -1,14 +1,15 @@
-package com.vu.androidbasicapp.ui
+package com.vu.androidbasicapp
 
 import android.os.Bundle
-import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import com.vu.androidbasicapp.R
+import androidx.fragment.app.Fragment
 
-
-class DashboardFragment : Fragment() {
+class ProfileFragment : Fragment() {
+    // TODO: Rename and change types of parameters
+    private var param1: String? = null
+    private var param2: String? = null
 
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
@@ -20,6 +21,6 @@ class DashboardFragment : Fragment() {
         savedInstanceState: Bundle?
     ): View? {
         // Inflate the layout for this fragment
-        return inflater.inflate(R.layout.fragment_dashboard, container, false)
+        return inflater.inflate(R.layout.fragment_profile, container, false)
     }
 }
