+++ app/src/main/java/com/vu/androidbasicapp/home/ui/HomeScreenFragment.kt
@@ -1,4 +1,4 @@
-package com.vu.androidbasicapp.ui.home
+package com.vu.androidbasicapp.home.ui
 
 import android.os.Bundle
 import android.view.LayoutInflater
