+++ app/src/main/java/com/vu/androidbasicapp/home/ui/HomeScreenViewModel.kt
@@ -1,11 +1,10 @@
-package com.vu.androidbasicapp.ui.home
+package com.vu.androidbasicapp.home.ui
 
 import android.util.Log
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.MutableStateFlow
-import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.launch
 
 class HomeScreenViewModel: ViewModel() {
