+++ app/src/main/res/layout/fragment_dashboard.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".ui.DashboardFragment">
+    tools:context=".DashboardFragment">
 
     <TextView
         android:id="@+id/screenTitle"
