+++ app/src/main/res/navigation/nav_graph.xml
@@ -7,17 +7,17 @@
 
     <fragment
         android:id="@+id/homeScreenFragment"
-        android:name="com.vu.androidbasicapp.ui.home.HomeScreenFragment"
+        android:name="com.vu.androidbasicapp.home.ui.HomeScreenFragment"
         android:label="fragment_home_screen"
         tools:layout="@layout/fragment_home_screen" />
     <fragment
         android:id="@+id/dashboardFragment"
-        android:name="com.vu.androidbasicapp.ui.DashboardFragment"
+        android:name="com.vu.androidbasicapp.DashboardFragment"
         android:label="fragment_dashboard"
         tools:layout="@layout/fragment_dashboard" />
     <fragment
         android:id="@+id/profileFragment"
-        android:name="com.vu.androidbasicapp.ui.ProfileFragment"
+        android:name="com.vu.androidbasicapp.ProfileFragment"
         android:label="fragment_profile"
         tools:layout="@layout/fragment_profile" />
 </navigation>
